requirements.txt
anyio==3.5.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
base58==2.1.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
ECPy==1.2.5
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.9
websockets==10.2
xrpl-py==1.4.0
